Allow girder nginx root to be configurable
Many deployments use Girder as a backend & administrative service, with a deployment-specific web app as the user facing front page. In these cases, the normal Girder app is served out of a non-root path and the custom app will be served from the root. This variable makes that configuration straightforward.